refactor(challenge): 나열형 거대 파일 2개 추가 분리 (IntroCareerStart, RecruitmentInfo)
프레젠테이션 컴포넌트 나열형 2개를 파일명 기반 하위폴더로 분리:
- ChallengeIntroCareerStart 352→143줄 (challenge-intro-career-start/)
- RecruitmentInfoSection 353→203줄 (recruitment-info-section/)
순수 이동/추출, 동작 변경 없음. tsc 0 / lint 통과.
