+++ apps/web/src/domain/program/challenge/challenge-view/ChallengeIntroCareerStart.tsx
@@ -1,11 +1,12 @@
-import CircularBox from '@/common/box/CircularBox';
 import Heading2 from '@/common/header/Heading2';
 import { challengeColors } from '@/domain/program/challenge/ChallengeView';
 import Description from '@/domain/program/program-detail/Description';
-import { twMerge } from '@/lib/twMerge';
 import { ChallengeType, challengeTypeSchema } from '@/schema';
 import { ChallengePoint } from '@/types/interface';
-import { ReactNode, useMemo } from 'react';
+import { useMemo } from 'react';
+import GeneralTrendSection from './challenge-intro-career-start/GeneralTrendSection';
+import QuestionBoxGrid from './challenge-intro-career-start/QuestionBoxGrid';
+import ResumeTrendSection from './challenge-intro-career-start/ResumeTrendSection';
 
 const QUESTION_BOXES = [
   `나의 경험을\n전략적으로\n서류에 녹일 수\n없을까?`,
@@ -35,200 +36,6 @@ const {
   PERSONAL_STATEMENT_LARGE_CORP,
   EXPERIENCE_SUMMARY,
 } = challengeTypeSchema.enum;
-
-const TrendItem = ({
-  number,
-  title,
-  description,
-  imageSrc,
-  imageAlt,
-}: {
-  number: number;
-  title: ReactNode;
-  description: ReactNode;
-  imageSrc: string;
-  imageAlt: string;
-}) => (
-  <div>
-    <div className="mb-8 md:mb-20 md:flex md:items-center md:justify-between">
-      <div className="md:flex md:gap-3">
-        <CircularBox className="bg-primary text-xsmall14 md:text-small20 mb-2 h-5 w-5 shrink-0 font-semibold md:mt-0.5 md:h-8 md:w-8">
-          {number}
-        </CircularBox>
-        <div>
-          <Title>{title}</Title>
-          <Paragraph>{description}</Paragraph>
-        </div>
-      </div>
-      <img src={imageSrc} alt={imageAlt} />
-    </div>
-  </div>
-);
-
-// 이력서 템플릿용 트렌드 섹션
-const ResumeTrendSection = ({ primaryColor }: { primaryColor: string }) => (
-  <div className="flex w-full max-w-[1000px] flex-col gap-y-[50px] px-5 py-[70px] md:gap-y-20 md:px-10 md:py-[120px] lg:px-0">
-    <div className="flex w-full flex-col gap-y-3 md:items-center">
-      <p
-        className="text-xsmall16 md:text-small20 font-bold"
-        style={{ color: primaryColor }}
-      >
-        서류 작성의 시작, 이력서
-      </p>
-      <Heading2>
-        {new Date().getFullYear()} 합격 이력서 트렌드는{' '}
-        <span className="text-primary">핵심 역량 강조</span>
-        <br />
-        지원한 직무에 Fit한 역량을 드러내야 해요.
-      </Heading2>
-    </div>
-    <div className="flex w-full flex-col gap-y-[60px] md:gap-y-[70px]">
-      <TrendItem
-        number={1}
-        title={
-          <>
-            경험 나열만 하는 이력서가 아닌
-            <br />
-            <span className="text-primary">직무 역량이 돋보이는</span> 이력서가
-            필요해요.
-          </>
-        }
-        description={
-          <>
-            지원한 직무의 역량을 정의 후 핵심 역량을 강조하며 <br />
-            스스로의 Fit함을 요약적으로 제시하며 시작해요.
-          </>
-        }
-        imageSrc="/images/challenge-trend-143.svg"
-        imageAlt="직무정의, 역량강조, 직무와의fit"
-      />
-      <TrendItem
-        number={2}
-        title={
-          <>
-            이력서의 핵심은 How!
-            <br />
-            <span className="text-primary">
-              내가 어떻게 일하는 사람인지 드러내야 해요.
-            </span>
-          </>
-        }
-        description={
-          <>
-            경험을 드러낼 때는 과업의 배경부터 솔루션 도출까지
-            <br />
-            과정에서 어떻게 일했는지 드러나는 이력서가 필요해요.
-          </>
-        }
-        imageSrc="/images/challenge-trend-2-143.svg"
-        imageAlt="이력서의 핵심은 HOW! 내가 어떻게 일하는 사람인지 드러내야 해요."
-      />
-      <TrendItem
-        number={3}
-        title={
-          <>
-            최대한 구체적으로 작성!
-            <br />
-            <span className="text-primary">경험이 구체적일수록</span> 지원자의
-            Fit을 쉽게 파악해요.
-          </>
-        }
-        description={
-          <>
-            구체성은 곧 차별화!
-            <br />
-            경험이 구체적일수록 다른 지원자와의 차별화가 가능해요.
-          </>
-        }
-        imageSrc="/images/challenge-trend-3.svg"
-        imageAlt="구체적인 경험이 곧 차별화입니다."
-      />
-    </div>
-  </div>
-);
-
-// 일반 트렌드 섹션
-const GeneralTrendSection = ({ primaryColor }: { primaryColor: string }) => (
-  <div className="flex w-full max-w-[1000px] flex-col gap-y-[50px] px-5 py-[70px] md:gap-y-20 md:px-10 md:py-[120px] lg:px-0">
-    <div className="flex w-full flex-col gap-y-3 md:items-center">
-      <p
-        className="text-xsmall16 md:text-small20 font-bold"
-        style={{ color: primaryColor }}
-      >
-        취업 성공 전략
-      </p>
-      <Heading2>
-        {new Date().getFullYear()} 채용 트렌드는{' '}
-        <span className="text-primary">직무 연관성</span>,{' '}
-        <br className="hidden md:block" />
-        나에 <br className="md:hidden" />
-        대한 이해를 직무와 결합시켜야 해요
-      </Heading2>
-    </div>
-    <div className="flex w-full flex-col gap-y-[60px] md:gap-y-[70px]">
-      <TrendItem
-        number={1}
-        title={
-          <>
-            스펙 나열하기는 그만!
-            <br className="hidden md:block" /> 나만의 경험에서{' '}
-            <br className="md:hidden" />
-            <span className="text-primary">차별화 포인트</span>부터 찾아야 해요
-          </>
-        }
-        description={
-          <>
-            직무에서 선호하는 K(지식) / S(스킬) / A(태도)에 맞춰 <br />
-            쌓아 온 경험을 재구성하면 경쟁력을 갖출 수 있어요
-          </>
-        }
-        imageSrc="/images/challenge-trend-1.svg"
-        imageAlt="K(지식), S(스킬), A(태도)"
-      />
-      <TrendItem
-        number={2}
-        title={
-          <>
-            뻔한 말은 그만!
-            <br className="hidden md:block" />{' '}
-            <span className="text-primary">나만의 컨셉</span>이 있어야{' '}
-            <br className="md:hidden" />더 보고 싶은 서류가 완성돼요
-          </>
-        }
-        description={
-          <>
-            흔한 키워드가 아닌 채용 공고에서 강조하는 역량 키워드에 <br />
-            맞춰 정리하면, 다른 지원자들과의 차이를 만들 수 있어요
-          </>
-        }
-        imageSrc="/images/challenge-trend-2.svg"
-        imageAlt="형식적이고 뻔한 내용이 아니라 역량 키워드에 맞춰 서류를 정리해야 합니다."
-      />
-    </div>
-  </div>
-);
-
-// 질문 박스 그리드 컴포넌트
-const QuestionBoxGrid = ({ boxes }: { boxes: string[] }) => (
-  <div className="relative flex w-full flex-col items-center">
-    <div className="grid w-full grid-cols-3 gap-2 md:grid-cols-4 md:gap-3">
-      {boxes.map((box, index) => (
-        <div
-          key={index}
-          className={twMerge(
-            'md:text-small20 h-20 whitespace-pre rounded-md bg-white px-2.5 py-3 text-[10px] font-semibold md:h-40 md:px-5 md:py-6',
-            box === '' && 'hidden md:invisible md:block',
-          )}
-        >
-          {box}
-        </div>
-      ))}
-    </div>
-    {/* grid 양 옆에 회색 그라데이션 박스 */}
-    <div className="from-neutral-90 absolute bottom-0 left-0 top-0 h-full w-40 bg-gradient-to-r to-transparent md:w-80" />
-    <div className="from-neutral-90 absolute bottom-0 right-0 top-0 h-full w-40 bg-gradient-to-l to-transparent md:w-80" />
-  </div>
-);
 
 const ChallengeIntroCareerStart = ({
   challengeTitle,
@@ -332,21 +139,5 @@ const ChallengeIntroCareerStart = ({
     </section>
   );
 };
-
-function Title({ children }: { children?: ReactNode }) {
-  return (
-    <span className="text-small18 text-neutral-0 md:text-medium22 font-semibold md:font-bold">
-      {children}
-    </span>
-  );
-}
-
-function Paragraph({ children }: { children?: ReactNode }) {
-  return (
-    <p className="text-xsmall14 text-neutral-45 md:text-small18 mb-5 mt-2.5">
-      {children}
-    </p>
-  );
-}
 
 export default ChallengeIntroCareerStart;
