+++ apps/web/src/domain/program/challenge/template-view/recruitment-info-section/types.ts
@@ -1,0 +1,7 @@
+export interface PriceInfo {
+  title: string;
+  originalPrice: number;
+  discountAmount: number;
+  description: string;
+  planType: 'BASIC' | 'STANDARD' | 'PREMIUM' | 'LIGHT';
+}
